Add new memes - infinite story and 'let me in'
'Infinite story', or in original 'The NeverEnding Story', is an
English-language West German fantasy film released in 1984,
based on the novel of the same name by Michael Ende, about a boy
who reads a magical book that tells a story of a young warrior
whose task is to stop a dark force called the Nothing from
engulfing a mystical world.

Meme 'infinite story' is a movie poster, and should be used to
mock on something too much longing.

The second meme, 'let me in', is a couple of screenshots from
American comedian Eric Andre's show. It should be used to show
some desperate wishes.